refactor: restructure mock data, split CTA components, and wire Cmd+K search
- Move mock data into `lib/mock` directory using factory patterns
- Extract `SidebarCTA` and `MobileCTA` into isolated components via `useBountyCTAState` hook
- Enhance `search-command` to query cached bounties and fuzzy-search projects and pages
- Verified no type or lint errors remaining
