+++ app/api/leaderboard/top/route.ts
@@ -1,15 +1,15 @@
-import { NextResponse } from 'next/server';
-import { getMockLeaderboard } from '@/lib/mock-leaderboard';
+import { NextResponse } from "next/server";
+import { getMockLeaderboard } from "@/lib/mock/leaderboard";
 
 export async function GET(request: Request) {
-    const { searchParams } = new URL(request.url);
-    const count = parseInt(searchParams.get('count') || '5');
+  const { searchParams } = new URL(request.url);
+  const count = parseInt(searchParams.get("count") || "5");
 
-    // Simulate network delay
-    await new Promise(resolve => setTimeout(resolve, 200));
+  // Simulate network delay
+  await new Promise((resolve) => setTimeout(resolve, 200));
 
-    // Get top N contributors
-    const { data } = getMockLeaderboard(1, count);
+  // Get top N contributors
+  const { data } = getMockLeaderboard(1, count);
 
-    return NextResponse.json(data);
+  return NextResponse.json(data);
 }
