+++ components/bounty-detail/types.ts
@@ -1,0 +1,4 @@
+import { BountyFieldsFragment } from "@/lib/graphql/generated";
+import type { Bounty } from "@/types/bounty";
+
+export type SidebarBounty = BountyFieldsFragment & Partial<Bounty>;
